requirements.txt
Flask
PyMySQL
Werkzeug
mysql-connector-python